remove lorri, use nixDirenvIntegration
haven't used lorri in quite some time now so removal was in order,
especially since it is run as a daemon which does use some resources
unnecessarily.

nixDirenvIntegration is really great, not in the least for flakes
support - a lot better with caching etc.

See: https://github.com/nix-community/nix-direnv