hosts: use yaml instead of toml despite some drawbacks
The main drawback is that Nix has no builtin support for yaml, it
does for toml however. This is usually solved by just calling out to
some command (yj in this case) to convert it to something Nix supports
natively (eg. toml or json).

Unfortunately agenix doesn't really allow for the above, however - by
overriding the nix attr input to agenix we can force it to use
nixUnstable which ofc supports flakes and use a different mechanism
for getting the needed data.

The reason for wanting yaml rather than toml is that I personally
think that yaml is more readable when you've got more than 20 lines of config.